+++ src/ska_oso_services/common/__init__.py
@@ -1,5 +1,11 @@
-from ska_db_oda.rest.fastapicontext import FastAPIContext
+from ska_db_oda.unitofwork.postgresunitofwork import PostgresUnitOfWork
 
-# use of this is deprecated - inject the
-# ska_db_oda.persistence.fastapicontext.UnitOfWork Dependency instead
-oda = FastAPIContext()
+
+class _ODA:
+    """Thin wrapper to keep the oda.uow() call pattern used in PHT APIs."""
+
+    def uow(self):
+        return PostgresUnitOfWork()
+
+
+oda = _ODA()