+++ src/ska_oso_services/validation/sbdefinition.py
@@ -1,6 +1,7 @@
 from ska_oso_pdm import SBDefinition, Target, TelescopeType, ValidationArrayAssembly
 from ska_oso_pdm.sb_definition import CSPConfiguration, ScanDefinition
 
+from ska_oso_services.validation.constraints import validate_constraints
 from ska_oso_services.validation.csp import validate_csp
 from ska_oso_services.validation.mccs import validate_mccs
 from ska_oso_services.validation.model import ValidationContext, ValidationIssue, validator
@@ -16,7 +17,7 @@ def validate_sbdefinition(
     Applies all relevant Validators to the SBDefinition elements,
     collecting all the results into a single list.
 
-    :param sbd: the full SBDefinition to validate
+    :param sbd_context: the full SBDefinition to validate
     :return: the collated ValidationIssues resulting from applying all the
                 SBDefinition Validators
     """
@@ -45,10 +46,30 @@ def validate_sbdefinition(
                 primary_entity=target,
                 source_jsonpath=f"$.targets.{index}",
                 telescope=sbd.telescope,
+                relevant_context={"constraints": sbd.observing_constraints},
                 array_assembly=validation_array_assembly,
             )
         )
     ]
+
+    # observing constraints can truly be optional in an SBD, so
+    # only validating if present
+
+    if sbd.observing_constraints is not None:
+        constraint_validation_results = validate_constraints(
+            ValidationContext(
+                primary_entity=sbd.observing_constraints,
+                source_jsonpath="$.observing_constraints",
+                relevant_context={
+                    "targets": sbd.targets,
+                    "scan_definitions": _get_scan_sequence(sbd, preserve_subarray_beams=True),
+                },
+                telescope=sbd.telescope,
+                array_assembly=validation_array_assembly,
+            )
+        )
+    else:
+        constraint_validation_results = []
 
     csp_validation_results = [
         issue
@@ -101,6 +122,7 @@ def validate_sbdefinition(
 
     return (
         target_validation_results
+        + constraint_validation_results
         + receptor_validation_results
         + csp_validation_results
         + scan_validation_results
@@ -125,12 +147,20 @@ def _lookup_csp_configuration_for_scan(
     )
 
 
-def _get_scan_sequence(sbd: SBDefinition) -> list[ScanDefinition]:
+def _get_scan_sequence(
+    sbd: SBDefinition, preserve_subarray_beams: bool = False
+) -> list[ScanDefinition] | list[list[ScanDefinition]]:
     if sbd.telescope == TelescopeType.SKA_MID:
         return sbd.dish_allocations.scan_sequence
 
-    return [
-        scan
-        for subarray_beam in sbd.mccs_allocation.subarray_beams
-        for scan in subarray_beam.scan_sequence
-    ]
+    elif preserve_subarray_beams is True:
+        return [
+            subarray_beam.scan_sequence for subarray_beam in sbd.mccs_allocation.subarray_beams
+        ]
+
+    else:
+        return [
+            scan
+            for subarray_beam in sbd.mccs_allocation.subarray_beams
+            for scan in subarray_beam.scan_sequence
+        ]