+++ tests/component/__init__.py
@@ -4,3 +4,4 @@ KUBE_NAMESPACE = os.getenv("KUBE_NAMESPACE", "ska-oso-services")
 OSO_SERVICES_URL = os.environ["OSO_SERVICES_URL"]
 ODT_URL = f"{OSO_SERVICES_URL}/odt"
 PHT_URL = f"{OSO_SERVICES_URL}/pht"
+ENGINEERING_URL = f"{OSO_SERVICES_URL}/engineering"