+++ tests/unit/validation/test_target.py
@@ -1,4 +1,6 @@
+# pylint: disable=no-member
 import pytest
+from astropy import units as u
 from ska_oso_pdm import (
     AltAzCoordinates,
     GalacticCoordinates,
@@ -7,11 +9,11 @@ from ska_oso_pdm import (
     ValidationArrayAssembly,
 )
 from ska_oso_pdm.builders.target_builder import LowTargetBuilder, MidTargetBuilder
+from ska_oso_pdm.sb_definition import AltitudeConstraint, ObservingConstraints
 
 from ska_oso_services.validation.model import ValidationContext, ValidationIssueType
 from ska_oso_services.validation.target import (
-    validate_low_elevation,
-    validate_mid_elevation,
+    validate_elevation,
     validate_single_target_pst_beams,
     validate_target,
 )
@@ -25,6 +27,7 @@ def test_full_target_validation_for_valid_target(telescope):
             primary_entity=LMC_TARGET,
             telescope=telescope,
             array_assembly=ValidationArrayAssembly.AA05,
+            relevant_context={"constraints": ObservingConstraints()},
         )
     )
     assert result == []
@@ -37,6 +40,7 @@ def test_full_target_validation_for_galactic_target(telescope):
             primary_entity=GALACTIC_TARGET,
             telescope=telescope,
             array_assembly=ValidationArrayAssembly.AA05,
+            relevant_context={"constraints": ObservingConstraints()},
         )
     )
     assert result == []
@@ -49,6 +53,7 @@ def test_full_target_validation_for_altaz_target(telescope):
             primary_entity=ALTAZ_TARGET,
             telescope=telescope,
             array_assembly=ValidationArrayAssembly.AA05,
+            relevant_context={"constraints": ObservingConstraints()},
         )
     )
     assert result == []
@@ -61,6 +66,7 @@ def test_full_target_validation_for_sso_target(telescope):
             primary_entity=SSO_TARGET,
             telescope=telescope,
             array_assembly=ValidationArrayAssembly.AA05,
+            relevant_context={"constraints": ObservingConstraints()},
         )
     )
     assert result[0].level == ValidationIssueType.WARNING
@@ -75,9 +81,10 @@ def test_mid_target_below_min_elevation():
             ),
         ),
         telescope=TelescopeType.SKA_MID,
+        relevant_context={"constraints": ObservingConstraints()},
     )
-    result = validate_mid_elevation(input_context)
-    assert result[0].message == "Source never rises above 15.0 degrees"
+    result = validate_elevation(input_context)
+    assert result[0].message == "Source never rises above the horizon"
 
 
 def test_low_target_below_horizon():
@@ -87,9 +94,10 @@ def test_low_target_below_horizon():
             reference_coordinate=ICRSCoordinates(ra_str="19:38:25.2890", dec_str="+66:48:52.915"),
         ),
         telescope=TelescopeType.SKA_LOW,
+        relevant_context={"constraints": ObservingConstraints()},
     )
 
-    result = validate_low_elevation(input_context)
+    result = validate_elevation(input_context)
     assert result[0].message == "Source never rises above the horizon"
 
 
@@ -100,9 +108,10 @@ def test_low_galactic_target_below_horizon():
             reference_coordinate=GalacticCoordinates(l=123.0, b=27.0),
         ),
         telescope=TelescopeType.SKA_LOW,
+        relevant_context={"constraints": ObservingConstraints()},
     )
 
-    result = validate_low_elevation(input_context)
+    result = validate_elevation(input_context)
     assert result[0].message == "Source never rises above the horizon"
 
 
@@ -113,29 +122,70 @@ def test_low_altaz_target_below_horizon():
             reference_coordinate=AltAzCoordinates(az=270.0, el=-45.0),
         ),
         telescope=TelescopeType.SKA_LOW,
+        relevant_context={"constraints": ObservingConstraints()},
     )
 
-    result = validate_low_elevation(input_context)
+    result = validate_elevation(input_context)
     assert result[0].message == "Source never rises above the horizon"
 
 
-def test_low_target_below_min_elevation():
+def test_low_target_below_osd_min_elevation():
     input_context = ValidationContext(
         primary_entity=LowTargetBuilder(
             name="47 Tuc",
             reference_coordinate=ICRSCoordinates(ra_str="00:24:05.3590", dec_str="-72:04:53.200"),
         ),
         telescope=TelescopeType.SKA_LOW,
+        relevant_context={"constraints": ObservingConstraints()},
     )
 
-    result = validate_low_elevation(input_context)
+    result = validate_elevation(input_context)
 
     assert len(result) == 1
     assert (
-        result[0].message == "Maximum elevation (44.74 degrees) is less than 45.0 degrees "
-        "- performance may be degraded"
+        result[0].message
+        == "Maximum elevation (44.74 degrees) is less than the limit (45.0 degrees) "
     )
+    assert result[0].level == ValidationIssueType.ERROR
+
+
+def test_low_target_below_osd_elevation_with_constraints():
+    input_context = ValidationContext(
+        primary_entity=LowTargetBuilder(
+            name="47 Tuc",
+            reference_coordinate=ICRSCoordinates(ra_str="00:24:05.3590", dec_str="-72:04:53.200"),
+        ),
+        telescope=TelescopeType.SKA_LOW,
+        relevant_context={
+            "constraints": ObservingConstraints(altitude=AltitudeConstraint(min=30.0 * u.deg))
+        },
+    )
+
+    result = validate_elevation(input_context)
+
+    assert len(result) == 1
     assert result[0].level == ValidationIssueType.WARNING
+    assert "less than 45 degrees" in result[0].message
+
+
+def test_low_target_below_45_degrees_gives_performance_warning():
+    input_context = ValidationContext(
+        primary_entity=LowTargetBuilder(
+            name="47 Tuc",
+            reference_coordinate=ICRSCoordinates(ra_str="00:24:05.3590", dec_str="-72:04:53.200"),
+        ),
+        telescope=TelescopeType.SKA_LOW,
+        relevant_context={
+            "constraints": ObservingConstraints(altitude=AltitudeConstraint(min=30.0 * u.deg))
+        },
+    )
+
+    result = validate_elevation(input_context)
+
+    assert result[0].level == ValidationIssueType.WARNING
+    assert result[0].message == (
+        "Maximum elevation (44.74 degrees) is less than 45 degrees - performance may be degraded"
+    )
 
 
 def test_target_with_pst_beams(
@@ -147,6 +197,7 @@ def test_target_with_pst_beams(
         primary_entity=sbd.targets[0],
         telescope=TelescopeType.SKA_LOW,
         array_assembly=ValidationArrayAssembly.AA05,
+        relevant_context={"constraints": ObservingConstraints()},
     )
 
     result = validate_single_target_pst_beams(input_context)