+++ CHANGELOG.md
@@ -4,8 +4,38 @@ Changelog
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-Unreleased
+14.6.0
+**********
+* New major version of ODA Updated to ODA v18.0.0
+* Uses the updated `ska_db_oda.common.uow` UnitOfWork for ODA requests.
+* [BUGFIX] Fixed issue with target validation on generated Low commissioning SBs
+ 
+14.5.0
+***********
+* [BUGFIX] Fixes error message for MID targets below the elevation range
+*  Improves validation of Solar System objects
+* [BUGFIX] Restore LOW targets below 45 degree elevation warning
+* Adds `/engineering` API path with functionality to create and update EBs. This is to be used in notebooks now the ODA API is removed
+
+14.4.1
 ************
+* [BUGFIX] Fixes inconsistency with target validation and scan validation
+* [BUGFIX] Fixes bug where SBD without scans fails validation
+
+14.4.0
+************
+* Adds PST and PSS parameters to the /configuration response for LOW
+* Adds further validation for multiple subarray beams
+* Adds a new POST /{prj_id}/{ob_id}/generateGSMSurveySBDefinitions endpoint that will generate SBDefinitions
+  from the pointings file (currently stored in this repo as well) and other inputs.
+* Adds quality_attribute_metrics to low configurations from OSD
+* [BUGFIX] fixing constraint validation that did not correctly take the scan duration into account
+* [BUGFIX] fixing constraint validation that did not correctly account for LST constraints that span midnight
+
+
+14.3.0
+************
+* Adding validation for the observing constraints
 
 14.2.2
 ************